task: moved instance list to flyouts
